delete useWordDetail hook
